Move position hash to crate root
Moved PositionHash to the root of position crate.
