+++ position/src/prelude.rs
@@ -1,4 +1,5 @@
 pub use crate::castling_rights::CastlingRights;
 pub use crate::indexing::{Color, File, NegDiag, Piece, PosDiag, Rank, Square};
 pub use crate::move_generation::{ChessMove, ChessMoveHint, ChessUnmove};
-pub use crate::position::{ParseFenError, Position, PositionHash};
+pub use crate::position::{ParseFenError, Position};
+pub use crate::position_hash::PositionHash;
